requirements.txt
streamlit
cohere
python-dotenv
rapidfuzz
pymupdf
reportlab
